Rework deployment-related stuff
The gist of this is that `mod_wsgi-express setup-server` is used to
create a persistent HTTPD config on the staging server instead of
running `mod_wsgi-express start-server` in the background.

This allows for easier control of the app server via `run mod-wsgi` like
so:

    run mod-wsgi :restart  # NOTE: The colon is necessary

which will run `/var/mod_wsgi/sites/campusmap/apachectl restart` on the
staging server.

This also adds a custom VirtualHost config instead of relying on
`httpdmulti`.